Add 3 rules, restructure packages
